+++ CHANGELOG.md
@@ -20,6 +20,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - add parsing and string conversion for Red enum(pr [#59])
 - add parsing and string conversion for White enum(pr [#60])
 - add new shades and implement parsing from string(pr [#61])
+- add NamedColour trait and implement for Black(pr [#63])
 
 ### Changed
 
@@ -151,6 +152,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 [#60]: https://github.com/jerus-org/named-colour/pull/60
 [#61]: https://github.com/jerus-org/named-colour/pull/61
 [#62]: https://github.com/jerus-org/named-colour/pull/62
+[#63]: https://github.com/jerus-org/named-colour/pull/63
 [Unreleased]: https://github.com/jerus-org/named-colour/compare/v0.2.0...HEAD
 [0.2.0]: https://github.com/jerus-org/named-colour/compare/v0.1.6...v0.2.0
 [0.1.6]: https://github.com/jerus-org/named-colour/compare/v0.1.5...v0.1.6